feat(#228): accelerometer move to settings
