--- remote/src/app/modals/app-remote-settings/app-remote-settings.scss
@@ -1,10 +1,0 @@
-ion-list {
-  p {
-    font-size: 0.8rem;
-  }
-
-  ion-range {
-    padding-top: 0;
-    padding-bottom: 0;
-  }
-}
