+++ remote/src/app/pages/app-settings/app-settings.scss
@@ -1,0 +1,12 @@
+app-settings {
+  ion-list {
+    p {
+      font-size: 0.8rem;
+    }
+
+    ion-range {
+      padding-top: 0;
+      padding-bottom: 0;
+    }
+  }
+}
