+++ remote/src/global/app.scss
@@ -35,3 +35,4 @@
 @import 'theme/menu';
 @import 'theme/toolbar';
 @import 'theme/list';
+@import 'theme/segment';
