+++ remote/src/global/theme/segment.scss
@@ -1,0 +1,5 @@
+body.dark {
+  ion-segment {
+    --ion-text-color-rgb: var(--ion-color-light-rgb);
+  }
+}
